// frontend/src/services/generated/models/languageDiagnostic.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Servidor de Tradução
 * API para gerenciamento de traduções.
 * OpenAPI spec version: 1.0
 */
import type { NamespaceDiagnostic } from './namespaceDiagnostic';

export interface LanguageDiagnostic {
  language: string;
  isBase: boolean;
  totalTerms: number;
  translatedTerms: number;
  missingTerms: number;
  translatedPercentage: number;
  namespaces: NamespaceDiagnostic[];
}