// frontend/src/services/generated/models/systemDiagnostic.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Servidor de Tradução
 * API para gerenciamento de traduções.
 * OpenAPI spec version: 1.0
 */
import type { EnvironmentDiagnostic } from './environmentDiagnostic';

export interface SystemDiagnostic {
  system: string;
  totalTerms: number;
  translatedTerms: number;
  missingTerms: number;
  translatedPercentage: number;
  environments: EnvironmentDiagnostic[];
}